requirements.txt
ipykernel # for jupyter notebook
pandas
numpy
sentence_transformers # for sentence embeddings
umap-learn # for dimensionality reduction
hdbscan # for clustering
bertopic # for topic modeling
blingfire # for sentence tokenization
matplotlib
seaborn
plotly
wordcloud
scikit-learn 
tqdm # for progress bar
google-generativeai
